Update main.py
The idea here is that prioritizing the distance variable prioritizes the search more so in the direction of the end goal. If there is no obstacle in the path, it will be the fastest way there, but if there is an obstacle, it will again grow outward while still prioritizing the direction of the end goal. Increasing the number (10) will only prioritize the direction more. Conversely, if you reduce the number to lets so 0, you will get undirectied search which is the same as dijkstra's algorithm. For a randomly generated map it is probably better to have a large number here but 10 works just as well.